Reviewed the code to compile with flag -Wextra
libnetlink.c is the only file that hasn't been reviewed because
it requires more of a overhaul.

+++ xip/utils.c
@@ -13,7 +13,7 @@ int force = 0;
 
 int matches(const char *cmd, const char *pattern)
 {
-	int len = strlen(cmd);
+	size_t len = strlen(cmd);
 	if (len > strlen(pattern))
 		return -1;
 	return memcmp(pattern, cmd, len);
